pyproject.toml
[project]
name = "agenda"
version = "0.1.0"
description = "Task manager for Telegram. Plan, organize, and complete."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.23.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
]
